requirements.txt
flask==3.0.3
matplotlib==3.10.0
pandas==2.2.3
scikit-learn==1.5.2
sqlalchemy==2.0.37
gunicorn
psycopg2-binary==2.9.10
python-dotenv==1.0.1